labs/02-langchain/demo-16-simple-router-chain-api/pyproject.toml
[project]
name = "demo-16-simple-router-chain-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.128.0",
	"uvicorn==0.40.0",
	"python-dotenv==1.2.1",
	"requests==2.32.5",
	"langchain-openai==1.1.7",
	"langchain==1.2.8"
]
